Implemented parse within DBCString trait and moved most parse functions to their objects
