+++ src/attributes.rs
@@ -9,6 +9,19 @@ use derive_getters::Getters;
 use crate::DBCString;
 use crate::message::MessageId;
 
+use crate::parser;
+
+use nom::{
+    branch::alt,
+    bytes::complete::{tag, take_till},
+    multi::{many_till},
+    sequence::preceded,
+    combinator::{map, opt},
+    character::complete::{self, line_ending, multispace0},
+    number::complete::double,
+    IResult,
+};
+
 #[derive(Clone, Debug, PartialEq, Getters)]
 #[cfg_attr(feature = "serde", derive(Serialize))]
 pub struct AttributeDefault {
@@ -22,6 +35,27 @@ impl DBCString for AttributeDefault {
             self.attribute_name,
             self.attribute_value.dbc_string(),
         )
+    }
+    
+    fn parse(s: &str) -> IResult<&str, Self>
+        where
+            Self: Sized {
+        let (s, _) = multispace0(s)?;
+        let (s, _) = tag("BA_DEF_DEF_")(s)?;
+        let (s, _) = parser::ms1(s)?;
+        let (s, attribute_name) = parser::char_string(s)?;
+        let (s, _) = parser::ms1(s)?;
+        let (s, attribute_value) = AttributeValue::parse(s)?;
+        let (s, _) = parser::semi_colon(s)?;
+        let (s, _) = line_ending(s)?;
+
+        Ok((
+            s,
+            AttributeDefault {
+                attribute_name: attribute_name.to_string(),
+                attribute_value,
+            },
+        ))
     }
 }
 
@@ -39,6 +73,33 @@ impl DBCString for AttributeValueForObject {
             self.attribute_value.dbc_string(),
         )
     }
+
+    fn parse(s: &str) -> IResult<&str, Self>
+        where
+            Self: Sized {
+        let (s, _) = multispace0(s)?;
+        let (s, _) = tag("BA_")(s)?;
+        let (s, _) = parser::ms1(s)?;
+        let (s, attribute_name) = parser::char_string(s)?;
+        let (s, _) = parser::ms1(s)?;
+        // let (s, attribute_value) = alt((
+        //     network_node_attribute_value,
+        //     message_definition_attribute_value,
+        //     signal_attribute_value,
+        //     env_variable_attribute_value,
+        //     raw_attribute_value,
+        // ))(s)?;
+        let (s, attribute_value) = AttributeValuedForObjectType::parse(s)?;
+        let (s, _) = parser::semi_colon(s)?;
+        let (s, _) = line_ending(s)?;
+        Ok((
+            s,
+            AttributeValueForObject {
+                attribute_name: attribute_name.to_string(),
+                attribute_value,
+            },
+        ))
+    }
 }
 
 #[derive(Clone, Debug, PartialEq)]
@@ -55,6 +116,49 @@ pub enum AttributeDefinition {
     Plain(String),
 }
 
+impl AttributeDefinition {
+    // TODO add properties
+    fn attribute_definition_node(s: &str) -> IResult<&str, AttributeDefinition> {
+        let (s, _) = tag("BU_")(s)?;
+        let (s, _) = parser::ms1(s)?;
+        let (s, node) = take_till(parser::is_semi_colon)(s)?;
+        Ok((s, AttributeDefinition::Node(node.to_string())))
+    }
+
+    // TODO add properties
+    fn attribute_definition_signal(s: &str) -> IResult<&str, AttributeDefinition> {
+        let (s, _) = tag("SG_")(s)?;
+        let (s, _) = parser::ms1(s)?;
+        let (s, signal) = take_till(parser::is_semi_colon)(s)?;
+        Ok((s, AttributeDefinition::Signal(signal.to_string())))
+    }
+
+    // TODO add properties
+    fn attribute_definition_environment_variable(s: &str) -> IResult<&str, AttributeDefinition> {
+        let (s, _) = tag("EV_")(s)?;
+        let (s, _) = parser::ms1(s)?;
+        let (s, env_var) = take_till(parser::is_semi_colon)(s)?;
+        Ok((
+            s,
+            AttributeDefinition::EnvironmentVariable(env_var.to_string()),
+        ))
+    }
+
+    // TODO add properties
+    fn attribute_definition_message(s: &str) -> IResult<&str, AttributeDefinition> {
+        let (s, _) = tag("BO_")(s)?;
+        let (s, _) = parser::ms1(s)?;
+        let (s, message) = take_till(parser::is_semi_colon)(s)?;
+        Ok((s, AttributeDefinition::Message(message.to_string())))
+    }
+
+    // TODO add properties
+    fn attribute_definition_plain(s: &str) -> IResult<&str, AttributeDefinition> {
+        let (s, plain) = take_till(parser::is_semi_colon)(s)?;
+        Ok((s, AttributeDefinition::Plain(plain.to_string())))
+    }
+}
+
 impl DBCString for AttributeDefinition {
     fn dbc_string(&self) -> String {
         return match self {
@@ -64,6 +168,25 @@ impl DBCString for AttributeDefinition {
             Self::EnvironmentVariable(ev) => format!("EV_ {};", ev),
             Self:: Plain(s) => format!("{};", s),
         }
+    }
+
+    fn parse(s: &str) -> IResult<&str, Self>
+        where
+            Self: Sized {
+        let (s, _) = multispace0(s)?;
+        let (s, _) = tag("BA_DEF_")(s)?;
+        let (s, _) = parser::ms1(s)?;
+        let (s, def) = alt((
+            Self::attribute_definition_node,
+            Self::attribute_definition_signal,
+            Self::attribute_definition_environment_variable,
+            Self::attribute_definition_message,
+            Self::attribute_definition_plain,
+        ))(s)?;
+    
+        let (s, _) = parser::semi_colon(s)?;
+        let (s, _) = line_ending(s)?;
+        Ok((s, def))
     }
 }
 
@@ -80,6 +203,47 @@ pub enum ValueDescription {
         env_var_name: String,
         value_descriptions: Vec<ValDescription>,
     },
+}
+
+impl ValueDescription {
+    fn value_description_for_signal(s: &str) -> IResult<&str, ValueDescription> {
+        let (s, _) = parser::ms0(s)?;
+        let (s, _) = tag("VAL_")(s)?;
+        let (s, _) = parser::ms1(s)?;
+        let (s, message_id) = MessageId::parse(s)?;
+        let (s, _) = parser::ms1(s)?;
+        let (s, signal_name) = parser::c_ident(s)?;
+        let (s, value_descriptions) = many_till(
+            preceded(parser::ms1, ValDescription::parse),
+            preceded(opt(parser::ms1), parser::semi_colon),
+        )(s)?;
+        Ok((
+            s,
+            ValueDescription::Signal {
+                message_id,
+                signal_name,
+                value_descriptions: value_descriptions.0,
+            },
+        ))
+    }
+
+    fn value_description_for_env_var(s: &str) -> IResult<&str, ValueDescription> {
+        let (s, _) = parser::ms0(s)?;
+        let (s, _) = tag("VAL_")(s)?;
+        let (s, _) = parser::ms1(s)?;
+        let (s, env_var_name) = parser::c_ident(s)?;
+        let (s, value_descriptions) = many_till(
+            preceded(parser::ms1, ValDescription::parse),
+            preceded(opt(parser::ms1), parser::semi_colon),
+        )(s)?;
+        Ok((
+            s,
+            ValueDescription::EnvironmentVariable {
+                env_var_name,
+                value_descriptions: value_descriptions.0,
+            },
+        ))
+    }
 }
 
 impl DBCString for ValueDescription {
@@ -110,6 +274,15 @@ impl DBCString for ValueDescription {
             }
         }
     }
+
+    fn parse(s: &str) -> IResult<&str, Self>
+        where
+            Self: Sized {
+        let (s, _) = multispace0(s)?;
+        let (s, vd) = alt((Self::value_description_for_signal, Self::value_description_for_env_var))(s)?;
+        let (s, _) = line_ending(s)?;
+        Ok((s, vd))
+    }
 }
 
 
@@ -128,6 +301,65 @@ pub enum AttributeValuedForObjectType {
     MessageDefinitionAttributeValue(MessageId, Option<AttributeValue>),
     SignalAttributeValue(MessageId, String, AttributeValue),
     EnvVariableAttributeValue(String, AttributeValue),
+}
+
+impl AttributeValuedForObjectType {
+    fn network_node_attribute_value(s: &str) -> IResult<&str, AttributeValuedForObjectType> {
+        let (s, _) = tag("BU_")(s)?;
+        let (s, _) = parser::ms1(s)?;
+        let (s, node_name) = parser::c_ident(s)?;
+        let (s, _) = parser::ms1(s)?;
+        let (s, value) = AttributeValue::parse(s)?;
+        Ok((
+            s,
+            AttributeValuedForObjectType::NetworkNodeAttributeValue(node_name, value),
+        ))
+    }
+
+    fn message_definition_attribute_value(s: &str) -> IResult<&str, AttributeValuedForObjectType> {
+        let (s, _) = tag("BO_")(s)?;
+        let (s, _) = parser::ms1(s)?;
+        let (s, message_id) = MessageId::parse(s)?;
+        let (s, _) = parser::ms1(s)?;
+        let (s, value) = opt(AttributeValue::parse)(s)?;
+        Ok((
+            s,
+            AttributeValuedForObjectType::MessageDefinitionAttributeValue(message_id, value),
+        ))
+    }
+
+    fn signal_attribute_value(s: &str) -> IResult<&str, AttributeValuedForObjectType> {
+        let (s, _) = tag("SG_")(s)?;
+        let (s, _) = parser::ms1(s)?;
+        let (s, message_id) = MessageId::parse(s)?;
+        let (s, _) = parser::ms1(s)?;
+        let (s, signal_name) = parser::c_ident(s)?;
+        let (s, _) = parser::ms1(s)?;
+        let (s, value) = AttributeValue::parse(s)?;
+        Ok((
+            s,
+            AttributeValuedForObjectType::SignalAttributeValue(message_id, signal_name, value),
+        ))
+    }
+
+    fn env_variable_attribute_value(s: &str) -> IResult<&str, AttributeValuedForObjectType> {
+        let (s, _) = tag("EV_")(s)?;
+        let (s, _) = parser::ms1(s)?;
+        let (s, env_var_name) = parser::c_ident(s)?;
+        let (s, _) = parser::ms1(s)?;
+        let (s, value) = AttributeValue::parse(s)?;
+        Ok((
+            s,
+            AttributeValuedForObjectType::EnvVariableAttributeValue(env_var_name, value),
+        ))
+    }
+
+    fn raw_attribute_value(s: &str) -> IResult<&str, AttributeValuedForObjectType> {
+        map(
+            AttributeValue::parse,
+            AttributeValuedForObjectType::RawAttributeValue,
+        )(s)
+    }
 }
 
 impl DBCString for AttributeValuedForObjectType {
@@ -154,6 +386,12 @@ impl DBCString for AttributeValuedForObjectType {
          },
         }
     }
+
+    fn parse(s: &str) -> IResult<&str, Self>
+        where
+        Self: Sized {
+
+    }
 }
 
 #[derive(Clone, Debug, PartialEq)]
@@ -175,6 +413,28 @@ pub enum AttributeValue {
     AttributeValueCharString(String),
 }
 
+impl AttributeValue {
+    #[allow(dead_code)]
+    fn attribute_value_uint64(s: &str) -> IResult<&str, AttributeValue> {
+        map(complete::u64, AttributeValue::AttributeValueU64)(s)
+    }
+
+    #[allow(dead_code)]
+    fn attribute_value_int64(s: &str) -> IResult<&str, AttributeValue> {
+        map(complete::i64, AttributeValue::AttributeValueI64)(s)
+    }
+
+    fn attribute_value_f64(s: &str) -> IResult<&str, AttributeValue> {
+        map(double, AttributeValue::AttributeValueF64)(s)
+    }
+
+    fn attribute_value_charstr(s: &str) -> IResult<&str, AttributeValue> {
+        map(parser::char_string, |x| {
+            AttributeValue::AttributeValueCharString(x.to_string())
+        })(s)
+    }
+}
+
 impl DBCString for AttributeValue {
     fn dbc_string(&self) -> String {
         return match self {
@@ -183,6 +443,17 @@ impl DBCString for AttributeValue {
             Self::AttributeValueF64(val) => val.to_string(),
             Self::AttributeValueCharString(val) => val.to_string(),
         }
+    }
+
+    fn parse(s: &str) -> IResult<&str, Self>
+        where
+            Self: Sized {
+        alt((
+            // Self::attribute_value_uint64,
+            // Self::attribute_value_int64,
+            Self::attribute_value_f64,
+            Self::attribute_value_charstr,
+        ))(s)
     }
 }
 
@@ -206,6 +477,21 @@ impl DBCString for ValDescription {
             self.a, self.b
         )
     }
+
+    fn parse(s: &str) -> nom::IResult<&str, Self>
+        where
+            Self: Sized {
+        let (s, a) = double(s)?;
+        let (s, _) = parser::ms1(s)?;
+        let (s, b) = parser::char_string(s)?;
+        Ok((
+            s,
+            ValDescription {
+                a,
+                b: b.to_string(),
+            },
+        ))
+    }
 }
 
 /// Global value table
@@ -227,5 +513,24 @@ impl DBCString for ValueTable {
                 .collect::<Vec<String>>()
                 .join(";")
         )
+    }
+
+    fn parse(s: &str) -> nom::IResult<&str, Self>
+        where
+            Self: Sized {
+        let (s, _) = multispace0(s)?;
+        let (s, _) = tag("VAL_TABLE_")(s)?;
+        let (s, _) = parser::ms1(s)?;
+        let (s, value_table_name) = parser::c_ident(s)?;
+        let (s, value_descriptions) =
+            many_till(preceded(parser::ms0, ValDescription::parse), preceded(parser::ms0, parser::semi_colon))(s)?;
+        let (s, _) = line_ending(s)?;
+        Ok((
+            s,
+            ValueTable {
+                value_table_name,
+                value_descriptions: value_descriptions.0,
+            },
+        ))
     }
 }
