+++ src/nodes.rs
@@ -9,6 +9,18 @@ use derive_getters::Getters;
 use crate::DBCString;
 use crate::message::MessageId;
 
+use crate::parser;
+use nom::{
+    branch::alt,
+    bytes::complete::{tag},
+    character::complete::{multispace0, char, line_ending},
+    combinator::{map, opt, value},
+    sequence::preceded,
+    character::complete::{self, space0},
+    multi::separated_list0,
+    IResult,
+};
+
 /// CAN network nodes, names must be unique
 #[derive(Clone, Debug, PartialEq)]
 #[cfg_attr(feature = "serde", derive(Serialize))]
@@ -20,6 +32,17 @@ impl DBCString for Node {
             self.0.clone().join(" ")
         )
     }
+
+    fn parse(s: &str) -> nom::IResult<&str, Self>
+        where
+            Self: Sized {
+        let (s, _) = multispace0(s)?;
+        let (s, _) = tag("BU_:")(s)?;
+        let (s, li) = opt(preceded(parser::ms1, separated_list0(parser::ms1, parser::c_ident)))(s)?;
+        let (s, _) = space0(s)?;
+        let (s, _) = line_ending(s)?;
+        Ok((s, Node(li.unwrap_or_default())))   
+    }
 }
 
 #[derive(Clone, Debug, PartialEq)]
@@ -29,12 +52,28 @@ pub enum AccessNode {
     AccessNodeName(String),
 }
 
+impl AccessNode {
+    fn access_node_vector_xxx(s: &str) -> IResult<&str, AccessNode> {
+        value(AccessNode::AccessNodeVectorXXX, tag("VECTOR_XXX"))(s)
+    }
+
+    fn access_node_name(s: &str) -> IResult<&str, AccessNode> {
+        map(parser::c_ident, AccessNode::AccessNodeName)(s)
+    }
+}
+
 impl DBCString for AccessNode {
     fn dbc_string(&self) -> String {
         return match self {
             Self::AccessNodeName(s) => s.to_string(),
             Self::AccessNodeVectorXXX => "Vector__XXX".to_string(),
         }
+    }
+
+    fn parse(s: &str) -> nom::IResult<&str, Self>
+        where
+            Self: Sized {
+        alt((Self::access_node_vector_xxx, Self::access_node_name))(s)
     }
 }
 
@@ -45,6 +84,23 @@ pub enum AccessType {
     DummyNodeVector1,
     DummyNodeVector2,
     DummyNodeVector3,
+}
+
+impl AccessType {
+    fn dummy_node_vector_0(s: &str) -> IResult<&str, AccessType> {
+        value(AccessType::DummyNodeVector0, char('0'))(s)
+    }
+
+    fn dummy_node_vector_1(s: &str) -> IResult<&str, AccessType> {
+        value(AccessType::DummyNodeVector1, char('1'))(s)
+    }
+
+    fn dummy_node_vector_2(s: &str) -> IResult<&str, AccessType> {
+        value(AccessType::DummyNodeVector2, char('2'))(s)
+    }
+    fn dummy_node_vector_3(s: &str) -> IResult<&str, AccessType> {
+        value(AccessType::DummyNodeVector3, char('3'))(s)
+    }
 }
 
 impl DBCString for AccessType {
@@ -58,6 +114,19 @@ impl DBCString for AccessType {
           }
         )
     }
+
+    fn parse(s: &str) -> IResult<&str, Self>
+        where
+            Self: Sized {
+        let (s, _) = tag("DUMMY_NODE_VECTOR")(s)?;
+        let (s, node) = alt((
+            Self::dummy_node_vector_0,
+            Self::dummy_node_vector_1,
+            Self::dummy_node_vector_2,
+            Self::dummy_node_vector_3,
+        ))(s)?;
+        Ok((s, node))
+    }
 }
 
 #[derive(Clone, Debug, PartialEq)]
@@ -69,12 +138,29 @@ pub enum Transmitter {
     VectorXXX,
 }
 
+impl Transmitter {
+
+    fn transmitter_vector_xxx(s: &str) -> IResult<&str, Transmitter> {
+        value(Transmitter::VectorXXX, tag("Vector__XXX"))(s)
+    }
+
+    fn transmitter_node_name(s: &str) -> IResult<&str, Transmitter> {
+        map(parser::c_ident, Transmitter::NodeName)(s)
+    }
+}
+
 impl DBCString for Transmitter {
     fn dbc_string(&self) -> String {
         return match self {
             Self::NodeName(s) => s.to_string(),
             Self::VectorXXX => "Vector__XXX".to_string(),
         }
+    }
+
+    fn parse(s: &str) -> IResult<&str, Self>
+        where
+            Self: Sized {
+        alt((Self::transmitter_vector_xxx, Self::transmitter_node_name))(s)
     }
 }
 
@@ -83,6 +169,12 @@ impl DBCString for Transmitter {
 pub struct MessageTransmitter {
     pub (crate) message_id: MessageId,
     pub (crate) transmitter: Vec<Transmitter>,
+}
+
+impl MessageTransmitter {
+    fn message_transmitters(s: &str) -> IResult<&str, Vec<Transmitter>> {
+        separated_list0(parser::comma, Transmitter::parse)(s)
+    }
 }
 
 impl DBCString for MessageTransmitter {
@@ -97,5 +189,27 @@ impl DBCString for MessageTransmitter {
             .join(","),
             // TODO determine if it will be a problem to kick out Vector__XXX if no transmitter is defined
         )
+    }
+
+    fn parse(s: &str) -> IResult<&str, Self>
+        where
+            Self: Sized {
+        let (s, _) = multispace0(s)?;
+        let (s, _) = tag("BO_TX_BU_")(s)?;
+        let (s, _) = parser::ms1(s)?;
+        let (s, message_id) = MessageId::parse(s)?;
+        let (s, _) = parser::ms1(s)?;
+        let (s, _) = parser::colon(s)?;
+        let (s, _) = parser::ms1(s)?;
+        let (s, transmitter) = Self::message_transmitters(s)?;
+        let (s, _) = parser::semi_colon(s)?;
+        let (s, _) = line_ending(s)?;
+        Ok((
+            s,
+            MessageTransmitter {
+                message_id,
+                transmitter,
+            },
+        ))
     }
 }
